requirements.txt
asgiref==3.6.0
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.5.0
whitenoise==6.4.0
Pillow==9.5.0
django-recaptcha==3.0.0
django-axes==6.0.1
django-csp
django-crispy-forms==2.0
crispy-bootstrap4==2022.1